web/requirements.txt
django==1.9.4
djangorestframework==3.2.5
djangorestframework-gis==0.8.2
django-allauth==0.24.1
django-rest-auth==0.7.0
--allow-external Pillow
six
requests
Pillow
psycopg2
oauth2client==4.0.0
google-api-python-client==1.5.5
paypalrestsdk
